backend/pyproject.toml
[tool.poetry]
name = "foodgram-project-react-fastapi"
version = "0.1.0"
description = ""
authors = ["Artem <artni96@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
fastapi = "^0.115.4"
pydantic = "^2.9.2"
alembic = "^1.13.3"
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.0"
fastapi-users = "^13.0.0"
fastapi-users-db-sqlalchemy = "^6.0.1"
asyncpg = "^0.30.0"
black = "^24.10.0"
pyjwt = "2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytest = "^8.3.3"
redis = "^5.2.0"
poethepoet = "^0.29.0"
pytest-asyncio = "^0.24.0"
pytest-dotenv = "^0.5.2"
pytest-order = "^1.3.0"
jinja2 = "^3.1.4"
pdfkit = "^1.0.0"
httpx = "^0.27.2"
datamodel-code-generator = "^0.26.3"
fastapi-cache2 = "^0.2.2"
fastapi-login = "^1.10.3"
pytz = "^2024.2"
wkhtmltopdf = "^0.2"
loguru = "^0.7.3"
ruff = "^0.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
upload_ingredients = "python src/utils/upload_ingredients.py"
upload_tags = "python src/utils/upload_tags.py"

[tool.ruff]
exclude = [
    ".env",
    ".env-example",
    ".env-test",
    "alembic.ini",
    ""
]